# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EventHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /checkin/scan:
    post:
      operationId: scanQr
      tags: [checkin]
      summary: Validate QR token and mark check-in/out (volunteer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QrScanInput"
      responses:
        "200":
          description: Check-in/out result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinResult"
        "400":
          description: Invalid / already checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /checkin/manual:
    post:
      operationId: manualCheckin
      tags: [checkin]
      summary: Manual check-in by name/email fallback (volunteer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualCheckinInput"
      responses:
        "200":
          description: Check-in result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinResult"
        "404":
          description: Attendee not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Volunteer Applications ────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    QrScanInput:
      type: object
      required: [qrToken, eventId]
      properties:
        qrToken:
          type: string
        eventId:
          type: integer
        station:
          type: string
    ManualCheckinInput:
      type: object
      required: [eventId]
      properties:
        eventId:
          type: integer
        email:
          type: string
        name:
          type: string
        station:
          type: string
    CheckinResult:
      type: object
      required: [success, action, attendeeName, message]
      properties:
        success:
          type: boolean
        action:
          type: string
          enum: [check_in, check_out]
        attendeeName:
          type: string
        attendeeEmail:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time

    # ─── Attendance Dashboard ─────────────────────────────────────────────────
...